chore: move hook to its own hook file
