+++ src/utils/notifications.ts
@@ -71,28 +71,6 @@ export const userEnabledNotification = () => {
   return false
 }
 
-export const useNotificationPermissionState = () => {
-  const [notificationPermissionGranted, setNotificationPermissionGranted] = useState(
-    userEnabledNotification()
-  )
-
-  // Can not use navigator.permissions.query({name: 'notifications'}) as it won't work on most
-  // mobile browsers
-
-  useEffect(() => {
-    const permissionInterval = setInterval(() => {
-      if (userEnabledNotification()) {
-        setNotificationPermissionGranted(true)
-        clearInterval(permissionInterval)
-      }
-    }, 100)
-
-    return () => clearInterval(permissionInterval)
-  }, [])
-
-  return notificationPermissionGranted
-}
-
 /*
  * Trigger notification dialogue if supported
  * Returns true if permissions were granted
